Bank: organize database layer
