+++ bank/bank.go
@@ -14,24 +14,6 @@ var (
 	ErrLockedFunds           = errors.New("locked funds")
 )
 
-type BalanceStatus int8
-
-const (
-	BalanceAvailable BalanceStatus = iota + 1
-	BalanceLocked
-)
-
-func (a BalanceStatus) String() string {
-	switch a {
-	case BalanceAvailable:
-		return "Available"
-	case BalanceLocked:
-		return "Locked"
-	default:
-		return "Unknown" // nolint:goconst
-	}
-}
-
 type TransactionStatus int8
 
 const (
@@ -109,4 +91,5 @@ type Service interface {
 	Deposit(address string, price types.FIL) (types.FIL, error)
 	Withdraw(address string, destination string, price types.FIL) (types.FIL, error)
 	Balance(address string) (types.FIL, error)
+	Authorize(address string) (string, error)
 }
