+++ bank/postgres/authorize.go
@@ -1,0 +1,5 @@
+package postgres
+
+func (s BankService) Authorize(address string) (string, error) {
+	return "", nil
+}
